backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
python-multipart>=0.0.6
pydantic>=2.4.2
markdown>=3.5
weasyprint>=60.1
pypandoc>=1.13
google-generativeai>=0.3.0
Pillow>=10.1.0
python-docx>=1.0.1
python-pptx>=0.6.21
aiohttp>=3.8.6

pyinstaller>=6.0.0